Consolidate all the metadata wrangler related code into one file.
