+++ bin/metadata_upload_coverage
@@ -5,7 +5,7 @@ import sys
 bin_dir = os.path.split(__file__)[0]
 package_dir = os.path.join(bin_dir, "..")
 sys.path.append(os.path.abspath(package_dir))
-from api.coverage import MetadataUploadCoverageProvider
+from api.metadata_wrangler import MetadataUploadCoverageProvider
 from core.scripts import RunCollectionCoverageProviderScript
 
 RunCollectionCoverageProviderScript(
