+++ bin/metadata_wrangler_collection_updates
@@ -6,5 +6,5 @@ bin_dir = os.path.split(__file__)[0]
 package_dir = os.path.join(bin_dir, "..")
 sys.path.append(os.path.abspath(package_dir))
 from core.scripts import RunCollectionMonitorScript
-from api.monitor import MWCollectionUpdateMonitor
+from api.metadata_wrangler import MWCollectionUpdateMonitor
 RunCollectionMonitorScript(MWCollectionUpdateMonitor).run()
